Move State from tauri client to coconut interface
